extension/python_agents/pyproject.toml
[project]
name = "vibe_coding_agents"
version = "0.2.0"
description = "AI Code Assistant Agents for Vibe Coding"
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    # DeepAgents (主要框架)
    "deepagents>=0.2.5",
    
    # LangChain 核心
    "langchain>=1.0.2,<2.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-openai>=1.0.0,<2.0.0",
    
    # 系统监控
    "psutil>=5.9.0",
    
    # 环境配置
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
strict = true
ignore_missing_imports = true
enable_error_code = ["deprecated"]
disallow_any_generics = false

[dependency-groups]
dev = [
    "debugpy>=1.8.17",
]

